Added a way to display multiple levels and items.
Added a Health Item, and the level class now accepts an array of Health items to be displayed along with the player in the level. Nothing happens when the player and the health pack interact currently. Janky if statements for showing health packs and the player at the same time.