Added Portals between Levels, and more things to do with Levels.
The portals between Levels are @ for the entrance and & for the exit. The exit & is purely cosmetic and functions as a blank space. However, the entrance portal @ moves the player to a specific x and y in another level. The @ and & are both yellow colored, but it is hard to see.